remove eos implementation files
