+++ singularity-eos/eos/eos.hpp
@@ -27,1233 +27,28 @@
 #include <singularity-eos/eos/eos_base.hpp>
 #include <singularity-eos/eos/eos_variant.hpp>
 
-#ifdef SPINER_USE_HDF
-#include <hdf5.h>
-#include <hdf5_hl.h>
-#include <singularity-eos/base/fast-math/logs.hpp>
-#include <singularity-eos/base/root-finding-1d/root_finding.hpp>
-#include <spiner/databox.hpp>
-#include <spiner/spiner_types.hpp>
-#endif // SPINER_USE_HDF
-
+// Base stuff
 #include <singularity-eos/base/constants.hpp>
 #include <singularity-eos/base/eos_error.hpp>
 #include <singularity-eos/base/variadic_utils.hpp>
+
+// EOS models
+#include <singularity-eos/eos/eos_davis.hpp>
+#include <singularity-eos/eos/eos_gruneisen.hpp>
+#include <singularity-eos/eos/eos_eospac.hpp>
+#include <singularity-eos/eos/eos_ideal.hpp>
+#include <singularity-eos/eos/eos_jwl.hpp>
+#include <singularity-eos/eos/eos_spiner.hpp>
+#include <singularity-eos/eos/eos_stellar_collapse.hpp>
+
+// Modifiers
 #include <singularity-eos/eos/modifiers/eos_unitsystem.hpp>
 #include <singularity-eos/eos/modifiers/ramps_eos.hpp>
 #include <singularity-eos/eos/modifiers/relativistic_eos.hpp>
 #include <singularity-eos/eos/modifiers/scaled_eos.hpp>
 #include <singularity-eos/eos/modifiers/shifted_eos.hpp>
 
-#ifdef SINGULARITY_USE_EOSPAC
-#include <eos_Interface.h>
-#endif
-
 namespace singularity {
-
-using namespace eos_base;
-
-class IdealGas : public EosBase<IdealGas> {
- public:
-  IdealGas() = default;
-  PORTABLE_INLINE_FUNCTION IdealGas(Real gm1, Real Cv)
-      : _Cv(Cv), _gm1(gm1), _rho0(_P0 / (_gm1 * _Cv * _T0)), _sie0(_Cv * _T0),
-        _bmod0((_gm1 + 1) * _gm1 * _rho0 * _Cv * _T0), _dpde0(_gm1 * _rho0),
-        _dvdt0(1. / (_rho0 * _T0)) {}
-
-  IdealGas GetOnDevice() { return *this; }
-  PORTABLE_FUNCTION Real TemperatureFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real InternalEnergyFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real PressureFromDensityTemperature(const Real rho,
-                                                        const Real temperature,
-                                                        Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real PressureFromDensityInternalEnergy(const Real rho, const Real sie,
-                                                           Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real SpecificHeatFromDensityTemperature(const Real rho,
-                                                            const Real temperature,
-                                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real SpecificHeatFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real BulkModulusFromDensityTemperature(const Real rho,
-                                                           const Real temperature,
-                                                           Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real BulkModulusFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real GruneisenParamFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real GruneisenParamFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION void FillEos(Real &rho, Real &temp, Real &energy, Real &press,
-                                 Real &cv, Real &bmod, const unsigned long output,
-                                 Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  void ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press, Real &cv,
-                              Real &bmod, Real &dpde, Real &dvdt,
-                              Real *lambda = nullptr) const;
-  // Generic functions provided by the base class. These contain e.g. the vector
-  // overloads that use the scalar versions declared here
-  SG_ADD_BASE_CLASS_USINGS(IdealGas)
-  PORTABLE_INLINE_FUNCTION
-  int nlambda() const noexcept { return 0; }
-  static constexpr unsigned long PreferredInput() { return _preferred_input; }
-  PORTABLE_FUNCTION void PrintParams() const {
-    printf("Ideal Gas Parameters:\nGamma = %g\nCv    = %g\n", _gm1 + 1.0, _Cv);
-  }
-  PORTABLE_FUNCTION void DensityEnergyFromPressureTemperature(const Real press,
-                                                              const Real temp,
-                                                              Real *lambda, Real &rho,
-                                                              Real &sie) const;
-  inline void Finalize() {}
-  static std::string EosType() { return std::string("IdealGas"); }
-
- private:
-  Real _Cv, _gm1;
-  // reference values
-  Real _rho0, _sie0, _bmod0, _dpde0, _dvdt0;
-  static constexpr const Real _T0 = ROOM_TEMPERATURE;
-  static constexpr const Real _P0 = ATMOSPHERIC_PRESSURE;
-  // static constexpr const char _eos_type[] = {"IdealGas"};
-  static constexpr const unsigned long _preferred_input =
-      thermalqs::density | thermalqs::specific_internal_energy;
-};
-
-// COMMENT: This is meant to be an implementation of the Steinberg version of
-// the Gruneisen EOS which should correspond to eostype(3) in xRAGE and
-// /[...]/eos/gruneisen in FLAG
-class Gruneisen : public EosBase<Gruneisen> {
- public:
-  Gruneisen() = default;
-  PORTABLE_INLINE_FUNCTION
-  Gruneisen(const Real C0, const Real s1, const Real s2, const Real s3, const Real G0,
-            const Real b, const Real rho0, const Real T0, const Real P0, const Real Cv,
-            const Real rho_max)
-      : _C0(C0), _s1(s1), _s2(s2), _s3(s3), _G0(G0), _b(b), _rho0(rho0), _T0(T0), _P0(P0),
-        _Cv(Cv), _rho_max(rho_max) {
-    // Warn user when provided rho_max is greater than the computed rho_max
-#ifndef NDEBUG
-    const Real computed_rho_max = ComputeRhoMax(s1, s2, s3, rho0);
-    if (rho_max > RHOMAX_SAFETY * computed_rho_max) {
-      printf(
-          "WARNING: Provided rho_max, %g, is greater than the computed rho_max of %g.\n",
-          rho_max, computed_rho_max);
-      printf("         States beyond %g g/cm^3 are unphysical (i.e. imaginary sound "
-             "speeds).\n",
-             computed_rho_max);
-    }
-#endif
-  }
-  // Constructor when rho_max isn't specified automatically determines _rho_max
-  PORTABLE_INLINE_FUNCTION
-  Gruneisen(const Real C0, const Real s1, const Real s2, const Real s3, const Real G0,
-            const Real b, const Real rho0, const Real T0, const Real P0, const Real Cv)
-      : _C0(C0), _s1(s1), _s2(s2), _s3(s3), _G0(G0), _b(b), _rho0(rho0), _T0(T0), _P0(P0),
-        _Cv(Cv), _rho_max(RHOMAX_SAFETY * ComputeRhoMax(s1, s2, s3, rho0)) {}
-  static PORTABLE_FUNCTION Real ComputeRhoMax(const Real s1, const Real s2, const Real s3,
-                                              const Real rho0);
-  Gruneisen GetOnDevice() { return *this; }
-  PORTABLE_FUNCTION Real TemperatureFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real InternalEnergyFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real PressureFromDensityTemperature(const Real rho,
-                                                        const Real temperature,
-                                                        Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real PressureFromDensityInternalEnergy(const Real rho, const Real sie,
-                                                           Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real SpecificHeatFromDensityTemperature(const Real rho,
-                                                            const Real temperatummmmmmre,
-                                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real SpecificHeatFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real BulkModulusFromDensityTemperature(const Real rho,
-                                                           const Real temperature,
-                                                           Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real BulkModulusFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real GruneisenParamFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real GruneisenParamFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION void FillEos(Real &rho, Real &temp, Real &energy, Real &press,
-                                 Real &cv, Real &bmod, const unsigned long output,
-                                 Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  void ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press, Real &cv,
-                              Real &bmod, Real &dpde, Real &dvdt,
-                              Real *lambda = nullptr) const;
-  // Generic functions provided by the base class. These contain e.g. the vector
-  // overloads that use the scalar versions declared here
-  SG_ADD_BASE_CLASS_USINGS(Gruneisen)
-  PORTABLE_INLINE_FUNCTION
-  int nlambda() const noexcept { return 0; }
-  static constexpr unsigned long PreferredInput() { return _preferred_input; }
-  PORTABLE_FUNCTION void PrintParams() const {
-    static constexpr char s1[]{"Gruneisen Params: "};
-    printf("%s C0:%e s1:%e s2:%e s3:%e\n  G0:%e b:%e rho0:%e T0:%e\n  P0:%eCv:%e "
-           "rho_max:%e\n",
-           s1, _C0, _s1, _s2, _s3, _G0, _b, _rho0, _T0, _P0, _Cv, _rho_max);
-  }
-  PORTABLE_FUNCTION void DensityEnergyFromPressureTemperature(const Real press,
-                                                              const Real temp,
-                                                              Real *lambda, Real &rho,
-                                                              Real &sie) const;
-  inline void Finalize() {}
-  static std::string EosType() { return std::string("Gruneisen"); }
-
- private:
-  Real _C0, _s1, _s2, _s3, _G0, _b, _rho0, _T0, _P0, _Cv, _rho_max;
-  // static constexpr const char _eos_type[] = {"Gruneisen"};
-  PORTABLE_FUNCTION
-  Real Gamma(const Real rho) const;
-  PORTABLE_FUNCTION
-  Real dPres_drho_e(const Real rho, const Real sie) const;
-  static constexpr const unsigned long _preferred_input =
-      thermalqs::density | thermalqs::specific_internal_energy;
-  // Scaling factor for density singularity (reference pressure blows up). Consistent with
-  // Pagosa implementation
-  static constexpr Real RHOMAX_SAFETY = 0.99;
-};
-
-// COMMENT: This is meant to be an implementation of the "standard" JWL as
-// implemented in xRAGE for eostype(1).  It does not include any energy shifting
-class JWL : public EosBase<JWL> {
- public:
-  JWL() = default;
-  PORTABLE_INLINE_FUNCTION JWL(const Real A, const Real B, const Real R1, const Real R2,
-                               const Real w, const Real rho0, const Real Cv)
-      : _A(A), _B(B), _R1(R1), _R2(R2), _w(w), _rho0(rho0), _Cv(Cv) {}
-  JWL GetOnDevice() { return *this; }
-  PORTABLE_FUNCTION Real TemperatureFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real InternalEnergyFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real PressureFromDensityTemperature(const Real rho,
-                                                        const Real temperature,
-                                                        Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real PressureFromDensityInternalEnergy(const Real rho, const Real sie,
-                                                           Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real SpecificHeatFromDensityTemperature(const Real rho,
-                                                            const Real temperature,
-                                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real SpecificHeatFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real BulkModulusFromDensityTemperature(const Real rho,
-                                                           const Real temperature,
-                                                           Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real BulkModulusFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real GruneisenParamFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real GruneisenParamFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION void FillEos(Real &rho, Real &temp, Real &energy, Real &press,
-                                 Real &cv, Real &bmod, const unsigned long output,
-                                 Real *lambda = nullptr) const;
-  // Generic functions provided by the base class. These contain e.g. the vector
-  // overloads that use the scalar versions declared here
-  SG_ADD_BASE_CLASS_USINGS(JWL)
-  PORTABLE_FUNCTION
-  PORTABLE_INLINE_FUNCTION
-  int nlambda() const noexcept { return 0; }
-  PORTABLE_FUNCTION
-  void ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press, Real &cv,
-                              Real &bmod, Real &dpde, Real &dvdt,
-                              Real *lambda = nullptr) const;
-  static constexpr unsigned long PreferredInput() { return _preferred_input; }
-  PORTABLE_FUNCTION void PrintParams() const {
-    static constexpr char s1[]{"JWL Params: "};
-    printf("%sA:%e B:%e R1: %e\nR2:%e w:%e rho0:%e\nCv:%e\n", s1, _A, _B, _R1, _R2, _w,
-           _rho0, _Cv);
-  }
-  PORTABLE_FUNCTION void DensityEnergyFromPressureTemperature(const Real press,
-                                                              const Real temp,
-                                                              Real *lambda, Real &rho,
-                                                              Real &sie) const;
-  inline void Finalize() {}
-  static std::string EosType() { return std::string("JWL"); }
-
- private:
-  Real _A, _B, _R1, _R2, _w, _rho0, _Cv;
-  PORTABLE_INLINE_FUNCTION Real ReferenceEnergy(const Real rho) const;
-  PORTABLE_INLINE_FUNCTION Real ReferencePressure(const Real rho) const;
-  // static constexpr const char _eos_type[] = "JWL";
-  static constexpr const unsigned long _preferred_input =
-      thermalqs::density | thermalqs::specific_internal_energy;
-};
-
-class DavisReactants : public EosBase<DavisReactants> {
- public:
-  DavisReactants() = default;
-  PORTABLE_INLINE_FUNCTION
-  DavisReactants(const Real rho0, const Real e0, const Real P0, const Real T0,
-                 const Real A, const Real B, const Real C, const Real G0, const Real Z,
-                 const Real alpha, const Real Cv0)
-      : _rho0(rho0), _e0(e0), _P0(P0), _T0(T0), _A(A), _B(B), _C(C), _G0(G0), _Z(Z),
-        _alpha(alpha), _Cv0(Cv0) {}
-  DavisReactants GetOnDevice() { return *this; }
-  PORTABLE_FUNCTION Real TemperatureFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real InternalEnergyFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real PressureFromDensityTemperature(const Real rho,
-                                                        const Real temperature,
-                                                        Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real PressureFromDensityInternalEnergy(const Real rho, const Real sie,
-                                                           Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real SpecificHeatFromDensityTemperature(const Real rho,
-                                                            const Real temperature,
-                                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real SpecificHeatFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real BulkModulusFromDensityTemperature(const Real rho,
-                                                           const Real temperature,
-                                                           Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real BulkModulusFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real GruneisenParamFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real GruneisenParamFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION void FillEos(Real &rho, Real &temp, Real &energy, Real &press,
-                                 Real &cv, Real &bmod, const unsigned long output,
-                                 Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  void ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press, Real &cv,
-                              Real &bmod, Real &dpde, Real &dvdt,
-                              Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION void DensityEnergyFromPressureTemperature(const Real press,
-                                                              const Real temp,
-                                                              Real *lambda, Real &rho,
-                                                              Real &sie) const;
-  // Generic functions provided by the base class. These contain e.g. the vector
-  // overloads that use the scalar versions declared here
-  SG_ADD_BASE_CLASS_USINGS(DavisReactants)
-  PORTABLE_INLINE_FUNCTION
-  int nlambda() const noexcept { return 0; }
-  static constexpr unsigned long PreferredInput() { return _preferred_input; }
-  PORTABLE_FUNCTION void PrintParams() const {
-    static constexpr char s1[]{"DavisReactants Params: "};
-    printf("%srho0:%e e0:%e P0:%e\nT0:%e A:%e B:%e\nC:%e G0:%e Z:%e\nalpha:%e "
-           "Cv0:%e\n",
-           s1, _rho0, _e0, _P0, _T0, _A, _B, _C, _G0, _Z, _alpha, _Cv0);
-  }
-  void inline Finalize() {}
-  static std::string EosType() { return std::string("DavisReactants"); }
-
- private:
-  Real _rho0, _e0, _P0, _T0, _A, _B, _C, _G0, _Z, _alpha, _Cv0;
-  // static constexpr const char _eos_type[] = "DavisReactants";
-  static constexpr unsigned long _preferred_input =
-      thermalqs::density | thermalqs::specific_internal_energy;
-  PORTABLE_INLINE_FUNCTION Real Ps(const Real rho) const;
-  PORTABLE_INLINE_FUNCTION Real Es(const Real rho) const;
-  PORTABLE_INLINE_FUNCTION Real Ts(const Real rho) const;
-  PORTABLE_INLINE_FUNCTION Real Gamma(const Real rho) const;
-};
-
-class DavisProducts : public EosBase<DavisProducts> {
- public:
-  DavisProducts() = default;
-  PORTABLE_INLINE_FUNCTION
-  DavisProducts(const Real a, const Real b, const Real k, const Real n, const Real vc,
-                const Real pc, const Real Cv, const Real E0)
-      : _a(a), _b(b), _k(k), _n(n), _vc(vc), _pc(pc), _Cv(Cv), _E0(E0) {}
-  DavisProducts GetOnDevice() { return *this; }
-  PORTABLE_FUNCTION Real TemperatureFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real InternalEnergyFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real PressureFromDensityTemperature(const Real rho,
-                                                        const Real temperature,
-                                                        Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real PressureFromDensityInternalEnergy(const Real rho, const Real sie,
-                                                           Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real SpecificHeatFromDensityTemperature(const Real rho,
-                                                            const Real temperature,
-                                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real SpecificHeatFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real BulkModulusFromDensityTemperature(const Real rho,
-                                                           const Real temperature,
-                                                           Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real BulkModulusFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real GruneisenParamFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real GruneisenParamFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION void DensityEnergyFromPressureTemperature(const Real press,
-                                                              const Real temp,
-                                                              Real *lambda, Real &rho,
-                                                              Real &sie) const;
-  PORTABLE_FUNCTION void FillEos(Real &rho, Real &temp, Real &energy, Real &press,
-                                 Real &cv, Real &bmod, const unsigned long output,
-                                 Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  void ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press, Real &cv,
-                              Real &bmod, Real &dpde, Real &dvdt,
-                              Real *lambda = nullptr) const;
-  // Generic functions provided by the base class. These contain e.g. the vector
-  // overloads that use the scalar versions declared here
-  SG_ADD_BASE_CLASS_USINGS(DavisProducts)
-  PORTABLE_INLINE_FUNCTION
-  int nlambda() const noexcept { return 0; }
-  static constexpr unsigned long PreferredInput() { return _preferred_input; }
-  PORTABLE_FUNCTION void PrintParams() const {
-    static constexpr char s1[]{"DavisProducts Params: "};
-    printf("%sa:%e b:%e k:%e\nn:%e vc:%e pc:%e\nCv:%e E0:%e\n", s1, _a, _b, _k, _n, _vc,
-           _pc, _Cv, _E0);
-  }
-  inline void Finalize() {}
-  static std::string EosType() { return std::string("DavisProducts"); }
-
- private:
-  Real _a, _b, _k, _n, _vc, _pc, _Cv, _E0;
-  // static constexpr const char _eos_type[] = "DavisProducts";
-  static constexpr const unsigned long _preferred_input =
-      thermalqs::density | thermalqs::specific_internal_energy;
-  PORTABLE_INLINE_FUNCTION Real F(const Real rho) const;
-  PORTABLE_INLINE_FUNCTION Real Ps(const Real rho) const;
-  PORTABLE_INLINE_FUNCTION Real Es(const Real rho) const;
-  PORTABLE_INLINE_FUNCTION Real Ts(const Real rho) const;
-  PORTABLE_INLINE_FUNCTION Real Gamma(const Real rho) const;
-};
-
-#ifdef SPINER_USE_HDF
-/*
-  Tables all have indep. variables log10(rho), log10(T)
-
-  Extrapolation strategy:
-  ------------------------------------------------------------
-  We use cold curve for low temperatures/energies
-  and assume ideal gas for large temperatures/energies.
-
-  For low densities, we floor the density. For high densities, we
-  we use log-linear extrapolation.
-*/
-class SpinerEOSDependsRhoT : public EosBase<SpinerEOSDependsRhoT> {
- public:
-  // A weakly typed index map for lambdas
-  struct Lambda {
-    enum Index { lRho = 0, lT = 1 };
-  };
-  // Generic functions provided by the base class. These contain
-  // e.g. the vector overloads that use the scalar versions declared
-  // here We explicitly list, rather than using the macro because we
-  // overload some methods.
-  using EosBase<SpinerEOSDependsRhoT>::TemperatureFromDensityInternalEnergy;
-  using EosBase<SpinerEOSDependsRhoT>::InternalEnergyFromDensityTemperature;
-  using EosBase<SpinerEOSDependsRhoT>::PressureFromDensityTemperature;
-  using EosBase<SpinerEOSDependsRhoT>::PressureFromDensityInternalEnergy;
-  using EosBase<SpinerEOSDependsRhoT>::SpecificHeatFromDensityTemperature;
-  using EosBase<SpinerEOSDependsRhoT>::SpecificHeatFromDensityInternalEnergy;
-  using EosBase<SpinerEOSDependsRhoT>::BulkModulusFromDensityTemperature;
-  using EosBase<SpinerEOSDependsRhoT>::BulkModulusFromDensityInternalEnergy;
-  using EosBase<SpinerEOSDependsRhoT>::GruneisenParamFromDensityTemperature;
-  using EosBase<SpinerEOSDependsRhoT>::GruneisenParamFromDensityInternalEnergy;
-  using EosBase<SpinerEOSDependsRhoT>::PTofRE;
-  using EosBase<SpinerEOSDependsRhoT>::FillEos;
-
-  SpinerEOSDependsRhoT(const std::string &filename, int matid,
-                       bool reproduciblity_mode = false);
-  SpinerEOSDependsRhoT(const std::string &filename, const std::string &materialName,
-                       bool reproducibility_mode = false);
-  PORTABLE_INLINE_FUNCTION
-  SpinerEOSDependsRhoT() : memoryStatus_(DataStatus::Deallocated) {}
-
-  SpinerEOSDependsRhoT GetOnDevice();
-
-  PORTABLE_FUNCTION
-  Real TemperatureFromDensityInternalEnergy(const Real rho, const Real sie,
-                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real InternalEnergyFromDensityTemperature(const Real rho, const Real temperature,
-                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real PressureFromDensityTemperature(const Real rho, const Real temperature,
-                                      Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real PressureFromDensityInternalEnergy(const Real rho, const Real sie,
-                                         Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real SpecificHeatFromDensityTemperature(const Real rho, const Real temperature,
-                                          Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real SpecificHeatFromDensityInternalEnergy(const Real rho, const Real sie,
-                                             Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real BulkModulusFromDensityTemperature(const Real rho, const Real temperature,
-                                         Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real BulkModulusFromDensityInternalEnergy(const Real rho, const Real sie,
-                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real GruneisenParamFromDensityTemperature(const Real rho, const Real temperature,
-                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real GruneisenParamFromDensityInternalEnergy(const Real rho, const Real sie,
-                                               Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  void DensityEnergyFromPressureTemperature(const Real press, const Real temp,
-                                            Real *lambda, Real &rho, Real &sie) const;
-  PORTABLE_FUNCTION
-  void FillEos(Real &rho, Real &temp, Real &energy, Real &press, Real &cv, Real &bmod,
-               const unsigned long output, Real *lambda = nullptr) const;
-
-  PORTABLE_FUNCTION
-  void ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press, Real &cv,
-                              Real &bmod, Real &dpde, Real &dvdt,
-                              Real *lambda = nullptr) const;
-
-  PORTABLE_FUNCTION
-  Real RhoPmin(const Real temp) const;
-
-  static constexpr unsigned long PreferredInput() { return _preferred_input; }
-  int matid() const { return matid_; }
-  PORTABLE_INLINE_FUNCTION Real lRhoOffset() const { return lRhoOffset_; }
-  PORTABLE_INLINE_FUNCTION Real lTOffset() const { return lTOffset_; }
-  PORTABLE_INLINE_FUNCTION Real rhoMin() const { return rho_(lRhoMin_); }
-  PORTABLE_INLINE_FUNCTION Real rhoMax() const { return rho_(lRhoMax_); }
-  PORTABLE_INLINE_FUNCTION Real TMin() const { return T_(lTMin_); }
-  PORTABLE_INLINE_FUNCTION Real TMax() const { return TMax_; }
-  PORTABLE_INLINE_FUNCTION void PrintParams() const {
-    static constexpr char s1[]{"SpinerEOS Parameters:"};
-    static constexpr char s2[]{"depends on log_10(rho) and log_10(temp)"};
-    static constexpr char s3[]{"EOS file   = "};
-    static constexpr char s4[]{"EOS mat ID = "};
-    static constexpr char s5[]{"EOS name   = "};
-    printf("%s\n\t%s\n\t%s\n\t%s%i\n\t%s\n", s1, s2, s3, s4, matid_, s5);
-    return;
-  }
-  PORTABLE_FORCEINLINE_FUNCTION Real MinimumDensity() const { return rhoMin(); }
-  PORTABLE_FORCEINLINE_FUNCTION Real MinimumTemperature() const { return T_(lTMin_); }
-  PORTABLE_INLINE_FUNCTION
-  int nlambda() const noexcept { return _n_lambda; }
-  PORTABLE_INLINE_FUNCTION
-  RootFinding1D::Status rootStatus() const { return status_; }
-  PORTABLE_INLINE_FUNCTION
-  TableStatus tableStatus() const { return whereAmI_; }
-  RootFinding1D::RootCounts counts;
-  void Finalize();
-  static std::string EosType() { return std::string("SpinerEOSDependsRhoT"); }
-
- private:
-  herr_t loadDataboxes_(const std::string &matid_str, hid_t file, hid_t lTGroup,
-                        hid_t coldGroup);
-  void fixBulkModulus_();
-  void setlTColdCrit_();
-
-  static PORTABLE_FORCEINLINE_FUNCTION Real toLog_(const Real x, const Real offset) {
-    // return std::log10(x + offset + EPS);
-    // return std::log10(std::abs(std::max(x,-offset) + offset)+EPS);
-    return FastMath::log10(std::abs(std::max(x, -offset) + offset) + EPS);
-  }
-  static PORTABLE_FORCEINLINE_FUNCTION Real fromLog_(const Real lx, const Real offset) {
-    return FastMath::pow10(lx) - offset;
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  lRho_(const Real rho) const noexcept {
-    Real out = toLog_(rho, lRhoOffset_);
-    return out;
-    // return out < lRhoMin_ ? lRhoMin_ : out;
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  lT_(const Real T) const noexcept {
-    return toLog_(T, lTOffset_);
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  rho_(const Real lRho) const noexcept {
-    Real rho = fromLog_(lRho, lRhoOffset_);
-    return rho < 0 ? 0 : rho;
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  T_(const Real lT) const noexcept {
-    return fromLog_(lT, lTOffset_);
-  }
-
-  PORTABLE_FUNCTION
-  Real lTFromlRhoSie_(const Real lRho, const Real sie, TableStatus &whereAmI,
-                      Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real lTFromlRhoP_(const Real lRho, const Real press, TableStatus &whereAmI,
-                    Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real lRhoFromPlT_(const Real P, const Real lT, TableStatus &whereAmI,
-                    Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  void getLogsRhoT_(const Real rho, const Real temperature, Real &lRho, Real &lT,
-                    Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real sieFromlRhoTlT_(const Real lRho, const Real T, const Real lT,
-                       const TableStatus &whereAmI) const;
-  PORTABLE_FUNCTION
-  Real PFromRholRhoTlT_(const Real rho, const Real lRho, const Real T, const Real lT,
-                        const TableStatus &whereAmI) const;
-  PORTABLE_FUNCTION
-  Real CvFromlRholT_(const Real lRho, const Real lT, const TableStatus &whereAmI) const;
-  PORTABLE_FUNCTION
-  Real bModFromRholRhoTlT_(const Real rho, const Real lRho, const Real T, const Real lT,
-                           const TableStatus &whereAmI) const;
-  PORTABLE_FUNCTION
-  TableStatus getLocDependsRhoSie_(const Real lRho, const Real sie) const;
-  PORTABLE_FUNCTION
-  TableStatus getLocDependsRhoT_(const Real lRho, const Real lT) const;
-  // PORTABLE_FUNCTION
-  // Real sieToColdInterval_(const Real lRho, const Real sie) const;
-
-  static constexpr const unsigned long _preferred_input =
-      thermalqs::density | thermalqs::temperature;
-  // static constexpr const char _eos_type[] {"SpinerEOSDependsRhoT"};
-  static constexpr const int numDataBoxes_ = 12;
-  Spiner::DataBox P_, sie_, bMod_, dPdRho_, dPdE_, dTdRho_, dTdE_, dEdRho_, dEdT_;
-  Spiner::DataBox PMax_, sielTMax_, dEdTMax_, gm1Max_;
-  Spiner::DataBox lTColdCrit_;
-  Spiner::DataBox PCold_, sieCold_, bModCold_;
-  Spiner::DataBox dPdRhoCold_, dPdECold_, dTdRhoCold_, dTdECold_, dEdTCold_;
-  Spiner::DataBox rho_at_pmin_;
-  int numRho_, numT_;
-  Real lRhoMin_, lRhoMax_, rhoMax_;
-  Real lRhoMinSearch_;
-  Real lTMin_, lTMax_, TMax_;
-  Real rhoNormal_, TNormal_, sieNormal_, PNormal_;
-  Real CvNormal_, bModNormal_, dPdENormal_, dVdTNormal_;
-  Real lRhoOffset_, lTOffset_; // offsets must be non-negative
-  int matid_;
-  bool reproducible_;
-  // whereAmI_ and status_ used only for reporting. They are not thread-safe.
-  mutable TableStatus whereAmI_ = TableStatus::OnTable;
-  mutable RootFinding1D::Status status_ = RootFinding1D::Status::SUCCESS;
-  static constexpr const Real ROOT_THRESH = 1e-14; // TODO: experiment
-  static constexpr const Real SOFT_THRESH = 1e-8;
-  DataStatus memoryStatus_ = DataStatus::Deallocated;
-  static constexpr const int _n_lambda = 2;
-  static constexpr const char *_lambda_names[2] = {"log(rho)", "log(T)"};
-};
-
-/*
-  TODO(JMM): Extrapolation Strategy
-  ----------------------------------
-  Currently the bottom of the table is the bound.
-  We do constant extrapolation off the bottom of the table
-  and linear extrapolation off the top.
-
-  Since the table is in log-log space, this means extrapolation off
-  the top is a power law. Extrapolation off the bottom is constant.
-  This worked for nubhlight. But it may or may not work here. We will
-  potentially need to revisit this.
-
-  The best solution might be a three-part EOS matched to the bottom of
-  the table, containing:
-  - A polytropic term
-  - A photon pressure term
-  - An ideal gas term
-  mitigated by Ye and (1-Ye) to control how important each term is.
- */
-class SpinerEOSDependsRhoSie : public EosBase<SpinerEOSDependsRhoSie> {
- public:
-  struct SP5Tables {
-    Spiner::DataBox P, bMod, dPdRho, dPdE, dTdRho, dTdE, dEdRho;
-  };
-  // Generic functions provided by the base class. These contain
-  // e.g. the vector overloads that use the scalar versions declared
-  // here We explicitly list, rather than using the macro because we
-  // overload some methods.
-  using EosBase<SpinerEOSDependsRhoSie>::TemperatureFromDensityInternalEnergy;
-  using EosBase<SpinerEOSDependsRhoSie>::InternalEnergyFromDensityTemperature;
-  using EosBase<SpinerEOSDependsRhoSie>::PressureFromDensityTemperature;
-  using EosBase<SpinerEOSDependsRhoSie>::PressureFromDensityInternalEnergy;
-  using EosBase<SpinerEOSDependsRhoSie>::SpecificHeatFromDensityTemperature;
-  using EosBase<SpinerEOSDependsRhoSie>::SpecificHeatFromDensityInternalEnergy;
-  using EosBase<SpinerEOSDependsRhoSie>::BulkModulusFromDensityTemperature;
-  using EosBase<SpinerEOSDependsRhoSie>::BulkModulusFromDensityInternalEnergy;
-  using EosBase<SpinerEOSDependsRhoSie>::GruneisenParamFromDensityTemperature;
-  using EosBase<SpinerEOSDependsRhoSie>::GruneisenParamFromDensityInternalEnergy;
-  using EosBase<SpinerEOSDependsRhoSie>::PTofRE;
-  using EosBase<SpinerEOSDependsRhoSie>::FillEos;
-
-  PORTABLE_INLINE_FUNCTION SpinerEOSDependsRhoSie()
-      : memoryStatus_(DataStatus::Deallocated) {}
-  SpinerEOSDependsRhoSie(const std::string &filename, int matid,
-                         bool reproducibility_mode = false);
-  SpinerEOSDependsRhoSie(const std::string &filename, const std::string &materialName,
-                         bool reproducibility_mode = false);
-  SpinerEOSDependsRhoSie GetOnDevice();
-
-  PORTABLE_FUNCTION
-  Real TemperatureFromDensityInternalEnergy(const Real rho, const Real sie,
-                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real InternalEnergyFromDensityTemperature(const Real rho, const Real T,
-                                            Real *lambda = nullptr) const;
-
-  PORTABLE_FUNCTION
-  Real PressureFromDensityTemperature(const Real rho, const Real T,
-                                      Real *lambda = nullptr) const;
-
-  PORTABLE_FUNCTION
-  Real PressureFromDensityInternalEnergy(const Real rho, const Real sie,
-                                         Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real SpecificHeatFromDensityTemperature(const Real rho, const Real T,
-                                          Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real SpecificHeatFromDensityInternalEnergy(const Real rho, const Real sie,
-                                             Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real BulkModulusFromDensityTemperature(const Real rho, const Real T,
-                                         Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real BulkModulusFromDensityInternalEnergy(const Real rho, const Real sie,
-                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real GruneisenParamFromDensityTemperature(const Real rho, const Real T,
-                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real GruneisenParamFromDensityInternalEnergy(const Real rho, const Real sie,
-                                               Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  void DensityEnergyFromPressureTemperature(const Real press, const Real temp,
-                                            Real *lambda, Real &rho, Real &sie) const;
-  PORTABLE_FUNCTION
-  void FillEos(Real &rho, Real &temp, Real &energy, Real &press, Real &cv, Real &bmod,
-               const unsigned long output, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  void ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press, Real &cv,
-                              Real &bmod, Real &dpde, Real &dvdt,
-                              Real *lambda = nullptr) const;
-
-  static constexpr unsigned long PreferredInput() { return _preferred_input; }
-  int matid() const { return matid_; }
-  PORTABLE_INLINE_FUNCTION Real lRhoOffset() const { return lRhoOffset_; }
-  PORTABLE_INLINE_FUNCTION Real lTOffset() const { return lTOffset_; }
-  PORTABLE_INLINE_FUNCTION Real lEOffset() const { return lEOffset_; }
-  PORTABLE_INLINE_FUNCTION Real rhoMin() const { return fromLog_(lRhoMin_, lRhoOffset_); }
-  PORTABLE_INLINE_FUNCTION Real rhoMax() const { return fromLog_(lRhoMax_, lRhoOffset_); }
-  PORTABLE_INLINE_FUNCTION Real TMin() const {
-    return fromLog_(T_.range(0).min(), lTOffset_);
-  }
-  PORTABLE_INLINE_FUNCTION Real TMax() const {
-    return fromLog_(T_.range(0).max(), lTOffset_);
-  }
-  PORTABLE_INLINE_FUNCTION Real sieMin() const {
-    return fromLog_(sie_.range(0).min(), lEOffset_);
-  }
-  PORTABLE_INLINE_FUNCTION Real sieMax() const {
-    return fromLog_(sie_.range(0).max(), lEOffset_);
-  }
-
-  PORTABLE_FORCEINLINE_FUNCTION Real MinimumDensity() const { return rhoMin(); }
-  PORTABLE_FORCEINLINE_FUNCTION Real MinimumTemperature() const { return TMin(); }
-
-  PORTABLE_INLINE_FUNCTION
-  int nlambda() const noexcept { return _n_lambda; }
-  PORTABLE_INLINE_FUNCTION void PrintParams() const {
-    static constexpr char s1[]{"SpinerEOS Parameters:"};
-    static constexpr char s2[]{"depends on log_10(rho) and log_10(sie)"};
-    static constexpr char s3[]{"EOS mat ID = "};
-    printf("%s\n\t%s\n\t%s%i\n", s1, s2, s3, matid_);
-    return;
-  }
-  PORTABLE_INLINE_FUNCTION
-  RootFinding1D::Status rootStatus() const { return status_; }
-  RootFinding1D::RootCounts counts;
-  static std::string EosType() { return std::string("SpinerEOSDependsRhoSie"); }
-  void Finalize();
-
- private:
-  herr_t loadDataboxes_(const std::string &matid_str, hid_t file, hid_t lTGroup,
-                        hid_t lEGroup);
-  void calcBMod_(SP5Tables &tables);
-
-  static PORTABLE_FORCEINLINE_FUNCTION Real toLog_(const Real x, const Real offset) {
-    // return std::log10(std::abs(std::max(x,-offset) + offset)+EPS);
-    return FastMath::log10(std::abs(std::max(x, -offset) + offset) + EPS);
-  }
-  static PORTABLE_FORCEINLINE_FUNCTION Real fromLog_(const Real lx, const Real offset) {
-    return FastMath::pow10(lx) - offset;
-  }
-  PORTABLE_FUNCTION
-  Real interpRhoT_(const Real rho, const Real T, const Spiner::DataBox &db,
-                   Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real interpRhoSie_(const Real rho, const Real sie, const Spiner::DataBox &db,
-                     Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real lRhoFromPlT_(const Real P, const Real lT, Real *lambda) const;
-
-  Spiner::DataBox sie_; // depends on (rho,T)
-  Spiner::DataBox T_;   // depends on (rho, sie)
-  SP5Tables dependsRhoT_;
-  SP5Tables dependsRhoSie_;
-  int numRho_;
-  Real rhoNormal_, TNormal_, sieNormal_, PNormal_;
-  Real CvNormal_, bModNormal_, dPdENormal_, dVdTNormal_;
-  Real lRhoMin_, lRhoMax_, rhoMax_;
-  Spiner::DataBox PlRhoMax_, dPdRhoMax_;
-
-  Real lRhoOffset_, lTOffset_, lEOffset_; // offsets must be non-negative
-
-  static constexpr unsigned long _preferred_input =
-      thermalqs::density | thermalqs::temperature;
-  // static constexpr const char _eos_type[] = "SpinerEOSDependsRhoSie";
-  int matid_;
-  bool reproducible_;
-  mutable RootFinding1D::Status status_;
-  static constexpr const int _n_lambda = 1;
-  static constexpr const char *_lambda_names[1] = {"log(rho)"};
-  DataStatus memoryStatus_ = DataStatus::Deallocated;
-};
-
-// Note the Stellar Collapse tables have units of:
-// 1. Ye (unitless)
-// 2. log(MeV) for temperature
-// 3. log(g/cm^3) for density.
-//
-// TODO(JMM): For now the bottom of the table is a floor and the top
-// is linear extrapolation in log-log space. We should reconsider this
-// and introduce extrapolation as needed.
-class StellarCollapse : public EosBase<StellarCollapse> {
- public:
-  // A weakly typed index map for lambdas
-  struct Lambda {
-    enum Index { Ye = 0, lT = 1 };
-  };
-
-  // Generic functions provided by the base class. These contain
-  // e.g. the vector overloads that use the scalar versions declared
-  // here We explicitly list, rather than using the macro because we
-  // overload some methods.
-  using EosBase<StellarCollapse>::TemperatureFromDensityInternalEnergy;
-  using EosBase<StellarCollapse>::InternalEnergyFromDensityTemperature;
-  using EosBase<StellarCollapse>::PressureFromDensityTemperature;
-  using EosBase<StellarCollapse>::PressureFromDensityInternalEnergy;
-  using EosBase<StellarCollapse>::SpecificHeatFromDensityTemperature;
-  using EosBase<StellarCollapse>::SpecificHeatFromDensityInternalEnergy;
-  using EosBase<StellarCollapse>::BulkModulusFromDensityTemperature;
-  using EosBase<StellarCollapse>::BulkModulusFromDensityInternalEnergy;
-  using EosBase<StellarCollapse>::GruneisenParamFromDensityTemperature;
-  using EosBase<StellarCollapse>::GruneisenParamFromDensityInternalEnergy;
-  using EosBase<StellarCollapse>::PTofRE;
-  using EosBase<StellarCollapse>::FillEos;
-
-  StellarCollapse(const std::string &filename, bool use_sp5 = false,
-                  bool filter_bmod = true);
-
-  // Saves to an SP5 file
-  void Save(const std::string &filename);
-
-  PORTABLE_INLINE_FUNCTION
-  StellarCollapse() : memoryStatus_(DataStatus::Deallocated) {}
-
-  StellarCollapse GetOnDevice();
-
-  PORTABLE_FUNCTION
-  Real TemperatureFromDensityInternalEnergy(const Real rho, const Real sie,
-                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real InternalEnergyFromDensityTemperature(const Real rho, const Real temperature,
-                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real PressureFromDensityTemperature(const Real rho, const Real temperature,
-                                      Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real PressureFromDensityInternalEnergy(const Real rho, const Real sie,
-                                         Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real SpecificHeatFromDensityTemperature(const Real rho, const Real temperature,
-                                          Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real SpecificHeatFromDensityInternalEnergy(const Real rho, const Real sie,
-                                             Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real BulkModulusFromDensityTemperature(const Real rho, const Real temperature,
-                                         Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real BulkModulusFromDensityInternalEnergy(const Real rho, const Real sie,
-                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real GruneisenParamFromDensityTemperature(const Real rho, const Real temperature,
-                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  Real GruneisenParamFromDensityInternalEnergy(const Real rho, const Real sie,
-                                               Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  void DensityEnergyFromPressureTemperature(const Real press, const Real temp,
-                                            Real *lambda, Real &rho, Real &sie) const;
-  PORTABLE_FUNCTION
-  void FillEos(Real &rho, Real &temp, Real &energy, Real &press, Real &cv, Real &bmod,
-               const unsigned long output, Real *lambda = nullptr) const;
-
-  PORTABLE_FUNCTION
-  void ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press, Real &cv,
-                              Real &bmod, Real &dpde, Real &dvdt,
-                              Real *lambda = nullptr) const;
-  // Generic functions provided by the base class. These contain e.g. the vector
-  // overloads that use the scalar versions declared here
-  static constexpr unsigned long PreferredInput() { return _preferred_input; }
-  PORTABLE_FORCEINLINE_FUNCTION Real lRhoOffset() const { return lRhoOffset_; }
-  PORTABLE_FORCEINLINE_FUNCTION Real lTOffset() const { return lTOffset_; }
-  PORTABLE_FORCEINLINE_FUNCTION Real lEOffset() const { return lEOffset_; }
-  PORTABLE_FORCEINLINE_FUNCTION Real lRhoMin() const { return lRhoMin_; }
-  PORTABLE_FORCEINLINE_FUNCTION Real lRhoMax() const { return lRhoMax_; }
-  PORTABLE_FORCEINLINE_FUNCTION Real rhoMin() const { return rho_(lRhoMin_); }
-  PORTABLE_FORCEINLINE_FUNCTION Real rhoMax() const { return rho_(lRhoMax_); }
-  PORTABLE_FORCEINLINE_FUNCTION Real lTMin() const { return lTMin_; }
-  PORTABLE_FORCEINLINE_FUNCTION Real lTMax() const { return lTMax_; }
-  PORTABLE_FORCEINLINE_FUNCTION Real TMin() const { return T_(lTMin_); }
-  PORTABLE_FORCEINLINE_FUNCTION Real TMax() const { return T_(lTMax_); }
-  PORTABLE_FORCEINLINE_FUNCTION Real YeMin() const { return YeMin_; }
-  PORTABLE_FORCEINLINE_FUNCTION Real YeMax() const { return YeMax_; }
-  PORTABLE_FORCEINLINE_FUNCTION Real sieMin() const { return sieMin_; }
-  PORTABLE_FORCEINLINE_FUNCTION Real sieMax() const { return sieMax_; }
-  PORTABLE_INLINE_FUNCTION void PrintParams() const {
-    printf("StellarCollapse parameters:\n"
-           "depends on log10(rho), log10(T), Ye\n"
-           "lrho bounds = %g, %g\n"
-           "lT bounds = %g, %g\n"
-           "Ye bounds = %g, %g\n",
-           lRhoMin_, lRhoMax_, lTMin_, lTMax_, YeMin_, YeMax_);
-    return;
-  }
-  PORTABLE_FORCEINLINE_FUNCTION Real MinimumDensity() const { return rhoMin(); }
-  PORTABLE_FORCEINLINE_FUNCTION Real MinimumTemperature() const { return TMin(); }
-  PORTABLE_INLINE_FUNCTION
-  int nlambda() const noexcept { return _n_lambda; }
-  inline RootFinding1D::Status rootStatus() const { return status_; }
-  RootFinding1D::RootCounts counts;
-  void Finalize();
-  static std::string EosType() { return std::string("StellarCollapse"); }
-
- private:
-  void LoadFromSP5File_(const std::string &filename);
-  void LoadFromStellarCollapseFile_(const std::string &filename);
-  int readSCInt_(const hid_t &file_id, const std::string &name);
-  void readBounds_(const hid_t &file_id, const std::string &name, int size, Real &lo,
-                   Real &hi);
-  void readSCDset_(const hid_t &file_id, const std::string &name, Spiner::DataBox &db);
-
-  void medianFilter_(Spiner::DataBox &db);
-  void medianFilter_(const Spiner::DataBox &in, Spiner::DataBox &out);
-  void fillMedianBuffer_(Real buffer[], int width, int iY, int iT, int irho,
-                         const Spiner::DataBox &tab) const;
-  Real findMedian_(Real buffer[], int size) const;
-  void computeBulkModulus_();
-  void computeColdAndHotCurves_();
-  void setNormalValues_();
-
-  PORTABLE_INLINE_FUNCTION __attribute__((always_inline)) void
-  checkLambda_(Real *lambda) const noexcept {
-    if (lambda == nullptr) {
-      EOS_ERROR("StellarCollapse: lambda must contain Ye and 1 space for caching.\n");
-    }
-  }
-
-  PORTABLE_INLINE_FUNCTION __attribute__((always_inline)) Real
-  toLog_(const Real x, const Real offset) const noexcept {
-    // StellarCollapse can't use fast logs, unless we re-grid onto the
-    // "fast log grid"
-    return std::log10(std::abs(std::max(x, -offset) + offset) + EPS);
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  fromLog_(const Real lx, const Real offset) const noexcept {
-    // StellarCollapse can't use fast logs, unless we re-grid onto the
-    // "fast log grid"
-    return std::pow(10., lx) - offset;
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  lRho_(const Real rho) const noexcept {
-    Real out = toLog_(rho, lRhoOffset_);
-    return out;
-    // return out < lRhoMin_ ? lRhoMin_ : out;
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  lT_(const Real T) const noexcept {
-    return toLog_(T, lTOffset_);
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  rho_(const Real lRho) const noexcept {
-    Real rho = fromLog_(lRho, lRhoOffset_);
-    return rho < 0 ? 0 : rho;
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  T_(const Real lT) const noexcept {
-    return fromLog_(lT, lTOffset_);
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  le2e_(const Real le) const noexcept {
-    return fromLog_(le, lEOffset_);
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  e2le_(const Real e) const noexcept {
-    return toLog_(e, lEOffset_);
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  lP2P_(const Real lP) const noexcept {
-    return fromLog_(lP, lPOffset_);
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  P2lP_(const Real P) const noexcept {
-    return toLog_(P, lPOffset_);
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  lB2B_(const Real lB) const noexcept {
-    return fromLog_(lB, lBOffset_);
-  }
-  PORTABLE_INLINE_FUNCTION Real __attribute__((always_inline))
-  B2lB_(const Real B) const noexcept {
-    return toLog_(B, lBOffset_);
-  }
-
-  PORTABLE_FUNCTION Real lTFromlRhoSie_(const Real lRho, const Real sie,
-                                        Real *lambda) const noexcept;
-  PORTABLE_INLINE_FUNCTION __attribute__((always_inline)) void
-  getLogsFromRhoT_(const Real rho, const Real temp, Real *lambda, Real &lRho, Real &lT,
-                   Real &Ye) const noexcept {
-    checkLambda_(lambda);
-    lRho = lRho_(rho);
-    lT = lT_(temp);
-    Ye = lambda[Lambda::Ye];
-    lambda[Lambda::lT] = lT;
-  }
-  PORTABLE_INLINE_FUNCTION __attribute__((always_inline)) void
-  getLogsFromRhoSie_(const Real rho, const Real sie, Real *lambda, Real &lRho, Real &lT,
-                     Real &Ye) const noexcept {
-    lRho = lRho_(rho);
-    lT = lTFromlRhoSie_(lRho, sie, lambda);
-    Ye = lambda[Lambda::Ye];
-    return;
-  }
-
-  static constexpr const unsigned long _preferred_input =
-      thermalqs::density | thermalqs::temperature;
-
-  // Dependent variables
-  Spiner::DataBox lP_, lE_, dPdRho_, dPdE_, dEdT_, lBMod_;
-  Spiner::DataBox entropy_; // kb/baryon
-  Spiner::DataBox Xa_;      // mass fraction of alpha particles
-  Spiner::DataBox Xh_;      // mass fraction of heavy ions
-  Spiner::DataBox Xn_;      // mass fraction of neutrons
-  Spiner::DataBox Xp_;      // mass fraction of protons
-  Spiner::DataBox Abar_;    // Average atomic mass
-  Spiner::DataBox Zbar_;    // Average atomic number
-  // Spiner::DataBox gamma_; // polytropic index. dlog(P)/dlog(rho).
-  // dTdRho_, dTdE_, dEdRho_, dEdT_;
-
-  // Bounds of dependent variables. Needed for root finding.
-  Spiner::DataBox eCold_, eHot_;
-
-  // Independent variable bounds
-  int numRho_, numT_, numYe_;
-  Real lRhoMin_, lRhoMax_;
-  Real lTMin_, lTMax_;
-  Real YeMin_, YeMax_;
-  Real sieMin_, sieMax_;
-
-  static constexpr Real MeV2GK_ = 11.604525006;
-  static constexpr Real GK2MeV_ = 1. / MeV2GK_;
-  static constexpr Real MeV2K_ = 1.e9 * MeV2GK_;
-  static constexpr Real K2MeV_ = 1. / MeV2K_;
-  static constexpr Real TNormal_ = 5 * GK2MeV_; // Threshold of NSE
-  static constexpr Real rhoNormal_ = 2.e12;     // 1./100'th of nuclear density
-  static constexpr Real YeNormal_ = 0.3517;     // Beta equilibrium value
-  Real sieNormal_, PNormal_;
-  Real CvNormal_, bModNormal_, dPdENormal_, dVdTNormal_;
-
-  // offsets must be non-negative
-  Real lEOffset_;
-  static constexpr Real lRhoOffset_ = 0.0; // TODO(JMM): Address if this ever changes
-  static constexpr Real lTOffset_ = 0.0;
-  static constexpr Real lPOffset_ = 0.0;
-  static constexpr Real lBOffset_ = 0.0;
-
-  // whereAmI_ and status_ used only for reporting. They are not thread-safe.
-  mutable RootFinding1D::Status status_ = RootFinding1D::Status::SUCCESS;
-  static constexpr const Real ROOT_THRESH = 1e-14; // TODO: experiment
-  DataStatus memoryStatus_ = DataStatus::Deallocated;
-  static constexpr const int _n_lambda = 2;
-  static constexpr const char *_lambda_names[] = {"Ye", "log(T)"};
-
-  // Stuff for median filter smoothing
-  static constexpr Real EPSSMOOTH = 10.0;
-  static constexpr int MF_W = 3;
-  static constexpr int MF_S = (2 * MF_W + 1) * (2 * MF_W + 1) * (2 * MF_W + 1);
-};
-#endif // SPINER_USE_HDF
-
-#ifdef SINGULARITY_USE_EOSPAC
-// Only really works in serial
-// Not really supported on device
-class EOSPAC : public EosBase<EOSPAC> {
- public:
-  EOSPAC() = default;
-  EOSPAC(int matid, bool invert_at_setup = false);
-  EOSPAC GetOnDevice() { return *this; }
-  PORTABLE_FUNCTION Real TemperatureFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real InternalEnergyFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real PressureFromDensityTemperature(const Real rho,
-                                                        const Real temperature,
-                                                        Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real PressureFromDensityInternalEnergy(const Real rho, const Real sie,
-                                                           Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real SpecificHeatFromDensityTemperature(const Real rho,
-                                                            const Real temperature,
-                                                            Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real SpecificHeatFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real BulkModulusFromDensityTemperature(const Real rho,
-                                                           const Real temperature,
-                                                           Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real BulkModulusFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real GruneisenParamFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION Real GruneisenParamFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION void FillEos(Real &rho, Real &temp, Real &energy, Real &press,
-                                 Real &cv, Real &bmod, const unsigned long output,
-                                 Real *lambda = nullptr) const;
-  PORTABLE_FUNCTION
-  void DensityEnergyFromPressureTemperature(const Real press, const Real temp,
-                                            Real *lambda, Real &rho, Real &sie) const;
-  PORTABLE_FUNCTION void ValuesAtReferenceState(Real &rho, Real &temp, Real &sie,
-                                                Real &press, Real &cv, Real &bmod,
-                                                Real &dpde, Real &dvdt,
-                                                Real *lambda = nullptr) const;
-
-  // Generic functions provided by the base class. These contain
-  // e.g. the vector overloads that use the scalar versions declared
-  // here We explicitly list, rather than using the macro because we
-  // overload some methods.
-  using EosBase<EOSPAC>::TemperatureFromDensityInternalEnergy;
-  using EosBase<EOSPAC>::InternalEnergyFromDensityTemperature;
-  using EosBase<EOSPAC>::PressureFromDensityTemperature;
-  using EosBase<EOSPAC>::PressureFromDensityInternalEnergy;
-  using EosBase<EOSPAC>::SpecificHeatFromDensityTemperature;
-  using EosBase<EOSPAC>::SpecificHeatFromDensityInternalEnergy;
-  using EosBase<EOSPAC>::BulkModulusFromDensityTemperature;
-  using EosBase<EOSPAC>::BulkModulusFromDensityInternalEnergy;
-  using EosBase<EOSPAC>::GruneisenParamFromDensityTemperature;
-  using EosBase<EOSPAC>::GruneisenParamFromDensityInternalEnergy;
-  using EosBase<EOSPAC>::PTofRE;
-  using EosBase<EOSPAC>::FillEos;
-
-  // TODO (JHP): Change EOSPAC vector implementations to be more performant
-  // template<typename RealIndexer, typename ConstRealIndexer, typename LambdaIndexer>
-  // inline
-  // void TemperatureFromDensityInternalEnergy(ConstRealIndexer &&rhos,
-  //                                           ConstRealIndexer &&sies,
-  //                                           RealIndexer &&temperatures,
-  //                                           const int num,
-  //                                           LambdaIndexer &&lambdas) const;
-  // template<typename RealIndexer, typename ConstRealIndexer, typename LambdaIndexer>
-  // inline
-  // void InternalEnergyFromDensityTemperature(ConstRealIndexer &&rhos,
-  //                                           ConstRealIndexer &&temperatures,
-  //                                           RealIndexer &&sies,
-  //                                           const int num,
-  //                                           LambdaIndexer &&lambdas) const;
-  // template<typename RealIndexer, typename ConstRealIndexer, typename LambdaIndexer>
-  // inline
-  // void PressureFromDensityTemperature(ConstRealIndexer &&rhos,
-  //                                     ConstRealIndexer &&temperatures,
-  //                                     RealIndexer &&pressures,
-  //                                     const int num,
-  //                                     LambdaIndexer &&lambdas) const;
-  // template<typename RealIndexer, typename ConstRealIndexer, typename LambdaIndexer>
-  // inline
-  // void PressureFromDensityInternalEnergy(ConstRealIndexer &&rhos,
-  //                                        ConstRealIndexer &&sies,
-  //                                        RealIndexer &&pressures,
-  //                                        const int num,
-  //                                        LambdaIndexer &&lambdas) const;
-  // template<typename RealIndexer, typename ConstRealIndexer, typename LambdaIndexer>
-  // inline
-  // void SpecificHeatFromDensityTemperature(ConstRealIndexer &&rhos,
-  //                                         ConstRealIndexer &&temperatures,
-  //                                         RealIndexer &&cvs,
-  //                                         const int num,
-  //                                         LambdaIndexer &&lambdas) const;
-  // template<typename RealIndexer, typename ConstRealIndexer, typename LambdaIndexer>
-  // inline
-  // void SpecificHeatFromDensityInternalEnergy(ConstRealIndexer &&rhos,
-  //                                            ConstRealIndexer &&sies,
-  //                                            RealIndexer &&cvs,
-  //                                            const int num,
-  //                                            LambdaIndexer &&lambdas) const;
-  // template<typename RealIndexer, typename ConstRealIndexer, typename LambdaIndexer>
-  // inline
-  // void BulkModulusFromDensityTemperature(ConstRealIndexer &&rhos,
-  //                                        ConstRealIndexer &&temperatures,
-  //                                        RealIndexer &&bmods,
-  //                                        const int num,
-  //                                        LambdaIndexer &&lambdas) const;
-  // template<typename RealIndexer, typename ConstRealIndexer, typename LambdaIndexer>
-  // inline
-  // void BulkModulusFromDensityInternalEnergy(ConstRealIndexer &&rhos,
-  //                                           ConstRealIndexer &&sies,
-  //                                           RealIndexer &&bmods,
-  //                                           const int num,
-  //                                           LambdaIndexer &&lambdas) const;
-
-  // template<typename RealIndexer, typename ConstRealIndexer, typename LambdaIndexer>
-  // inline
-  // void GruneisenParamFromDensityTemperature(ConstRealIndexer &&rhos,
-  //                                           ConstRealIndexer &&temperatures,
-  //                                           RealIndexer &&gm1s,
-  //                                           const int num,
-  //                                           LambdaIndexer &&lambdas) const;
-
-  // template<typename RealIndexer, typename ConstRealIndexer, typename LambdaIndexer>
-  // inline
-  // void GruneisenParamFromDensityInternalEnergy(ConstRealIndexer &&rhos,
-  //                                              ConstRealIndexer &&sies,
-  //                                              RealIndexer &&gm1s,
-  //                                              const int num,
-  //                                              LambdaIndexer &&lambdas) const;
-
-  // template<typename RealIndexer, typename LambdaIndexer>
-  // inline
-  // void FillEos(RealIndexer &&rhos, RealIndexer &&temps, RealIndexer &&energies,
-  //              RealIndexer &&presses, RealIndexer &&cvs, RealIndexer &&bmods,
-  //              const int num, const unsigned long output,
-  //              LambdaIndexer &&lambdas) const;
-  static constexpr unsigned long PreferredInput() { return _preferred_input; }
-  PORTABLE_INLINE_FUNCTION int nlambda() const noexcept { return 0; }
-  inline void Finalize() {}
-  static std::string EosType() { return std::string("EOSPAC"); }
-  inline void PrintParams() const { printf("EOSPAC parameters:\nmatid = %s\n", matid_); }
-  PORTABLE_FORCEINLINE_FUNCTION Real MinimumDensity() const { return rho_min_; }
-  PORTABLE_FORCEINLINE_FUNCTION Real MinimumTemperature() const { return temp_min_; }
-
- private:
-  static constexpr const unsigned long _preferred_input =
-      thermalqs::density | thermalqs::temperature;
-  int matid_;
-  static constexpr int NT = 5;
-  EOS_INTEGER PofRT_table_;
-  EOS_INTEGER TofRE_table_;
-  EOS_INTEGER EofRT_table_;
-  EOS_INTEGER RofPT_table_;
-  EOS_INTEGER TofRP_table_;
-  // EOS_INTEGER PofRE_table_;
-  EOS_INTEGER tablehandle[NT];
-  EOS_INTEGER EOS_Info_table_;
-  static constexpr Real temp_ref_ = 293;
-  Real rho_ref_ = 1;
-  Real sie_ref_ = 1;
-  Real press_ref_ = 1;
-  Real cv_ref_ = 1;
-  Real bmod_ref_ = 1;
-  Real dpde_ref_ = 1;
-  Real dvdt_ref_ = 1;
-  Real rho_min_ = 0;
-  Real temp_min_ = 0;
-};
-#endif // SINGULARITY_USE_EOSPAC
 
 // recreate variadic list
 template <typename... Ts>
