+++ singularity-eos/eos/eos_gruneisen.hpp
@@ -12,14 +12,116 @@
 // publicly and display publicly, and to permit others to do so.
 //------------------------------------------------------------------------------
 
+#ifndef _SINGULARITY_EOS_EOS_GRUNEISEN_HPP_
+#define _SINGULARITY_EOS_EOS_GRUNEISEN_HPP_
+
+#include <cmath>
+#include <cstdio>
+
 #include <limits>
 
 #include <singularity-eos/base/constants.hpp>
 #include <singularity-eos/base/root-finding-1d/root_finding.hpp>
-#include <singularity-eos/eos/eos.hpp>
+#include <singularity-eos/eos/eos_base.hpp>
 
 namespace singularity {
 
+using namespace eos_base;
+
+// COMMENT: This is meant to be an implementation of the Steinberg version of
+// the Gruneisen EOS which should correspond to eostype(3) in xRAGE and
+// /[...]/eos/gruneisen in FLAG
+class Gruneisen : public EosBase<Gruneisen> {
+ public:
+  Gruneisen() = default;
+  PORTABLE_INLINE_FUNCTION
+  Gruneisen(const Real C0, const Real s1, const Real s2, const Real s3, const Real G0,
+            const Real b, const Real rho0, const Real T0, const Real P0, const Real Cv,
+            const Real rho_max)
+      : _C0(C0), _s1(s1), _s2(s2), _s3(s3), _G0(G0), _b(b), _rho0(rho0), _T0(T0), _P0(P0),
+        _Cv(Cv), _rho_max(rho_max) {
+    // Warn user when provided rho_max is greater than the computed rho_max
+#ifndef NDEBUG
+    const Real computed_rho_max = ComputeRhoMax(s1, s2, s3, rho0);
+    if (rho_max > RHOMAX_SAFETY * computed_rho_max) {
+      printf(
+          "WARNING: Provided rho_max, %g, is greater than the computed rho_max of %g.\n",
+          rho_max, computed_rho_max);
+      printf("         States beyond %g g/cm^3 are unphysical (i.e. imaginary sound "
+             "speeds).\n",
+             computed_rho_max);
+    }
+#endif
+  }
+  // Constructor when rho_max isn't specified automatically determines _rho_max
+  PORTABLE_INLINE_FUNCTION
+  Gruneisen(const Real C0, const Real s1, const Real s2, const Real s3, const Real G0,
+            const Real b, const Real rho0, const Real T0, const Real P0, const Real Cv)
+      : _C0(C0), _s1(s1), _s2(s2), _s3(s3), _G0(G0), _b(b), _rho0(rho0), _T0(T0), _P0(P0),
+        _Cv(Cv), _rho_max(RHOMAX_SAFETY * ComputeRhoMax(s1, s2, s3, rho0)) {}
+  static PORTABLE_INLINE_FUNCTION Real ComputeRhoMax(const Real s1, const Real s2,
+                                                     const Real s3, const Real rho0);
+  Gruneisen GetOnDevice() { return *this; }
+  PORTABLE_INLINE_FUNCTION Real TemperatureFromDensityInternalEnergy(
+      const Real rho, const Real sie, Real *lambda = nullptr) const;
+  PORTABLE_INLINE_FUNCTION Real InternalEnergyFromDensityTemperature(
+      const Real rho, const Real temperature, Real *lambda = nullptr) const;
+  PORTABLE_INLINE_FUNCTION Real PressureFromDensityTemperature(
+      const Real rho, const Real temperature, Real *lambda = nullptr) const;
+  PORTABLE_INLINE_FUNCTION Real PressureFromDensityInternalEnergy(
+      const Real rho, const Real sie, Real *lambda = nullptr) const;
+  PORTABLE_INLINE_FUNCTION Real SpecificHeatFromDensityTemperature(
+      const Real rho, const Real temperatummmmmmre, Real *lambda = nullptr) const;
+  PORTABLE_INLINE_FUNCTION Real SpecificHeatFromDensityInternalEnergy(
+      const Real rho, const Real sie, Real *lambda = nullptr) const;
+  PORTABLE_INLINE_FUNCTION Real BulkModulusFromDensityTemperature(
+      const Real rho, const Real temperature, Real *lambda = nullptr) const;
+  PORTABLE_INLINE_FUNCTION Real BulkModulusFromDensityInternalEnergy(
+      const Real rho, const Real sie, Real *lambda = nullptr) const;
+  PORTABLE_INLINE_FUNCTION Real GruneisenParamFromDensityTemperature(
+      const Real rho, const Real temperature, Real *lambda = nullptr) const;
+  PORTABLE_INLINE_FUNCTION Real GruneisenParamFromDensityInternalEnergy(
+      const Real rho, const Real sie, Real *lambda = nullptr) const;
+  PORTABLE_INLINE_FUNCTION void FillEos(Real &rho, Real &temp, Real &energy, Real &press,
+                                        Real &cv, Real &bmod, const unsigned long output,
+                                        Real *lambda = nullptr) const;
+  PORTABLE_INLINE_FUNCTION
+  void ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press, Real &cv,
+                              Real &bmod, Real &dpde, Real &dvdt,
+                              Real *lambda = nullptr) const;
+  // Generic functions provided by the base class. These contain e.g. the vector
+  // overloads that use the scalar versions declared here
+  SG_ADD_BASE_CLASS_USINGS(Gruneisen)
+  PORTABLE_INLINE_FUNCTION
+  int nlambda() const noexcept { return 0; }
+  static constexpr unsigned long PreferredInput() { return _preferred_input; }
+  PORTABLE_INLINE_FUNCTION void PrintParams() const {
+    static constexpr char s1[]{"Gruneisen Params: "};
+    printf("%s C0:%e s1:%e s2:%e s3:%e\n  G0:%e b:%e rho0:%e T0:%e\n  P0:%eCv:%e "
+           "rho_max:%e\n",
+           s1, _C0, _s1, _s2, _s3, _G0, _b, _rho0, _T0, _P0, _Cv, _rho_max);
+  }
+  PORTABLE_INLINE_FUNCTION void
+  DensityEnergyFromPressureTemperature(const Real press, const Real temp, Real *lambda,
+                                       Real &rho, Real &sie) const;
+  inline void Finalize() {}
+  static std::string EosType() { return std::string("Gruneisen"); }
+
+ private:
+  Real _C0, _s1, _s2, _s3, _G0, _b, _rho0, _T0, _P0, _Cv, _rho_max;
+  // static constexpr const char _eos_type[] = {"Gruneisen"};
+  PORTABLE_INLINE_FUNCTION
+  Real Gamma(const Real rho) const;
+  PORTABLE_INLINE_FUNCTION
+  Real dPres_drho_e(const Real rho, const Real sie) const;
+  static constexpr const unsigned long _preferred_input =
+      thermalqs::density | thermalqs::specific_internal_energy;
+  // Scaling factor for density singularity (reference pressure blows up). Consistent with
+  // Pagosa implementation
+  static constexpr Real RHOMAX_SAFETY = 0.99;
+};
+
+namespace gruneisen_utils {
 PORTABLE_INLINE_FUNCTION Real square(const Real x) { return x * x; }
 PORTABLE_INLINE_FUNCTION Real cube(const Real x) { return x * x * x; }
 PORTABLE_INLINE_FUNCTION Real find_min_bounded_val(const Real val1, const Real val2,
@@ -42,12 +144,15 @@ PORTABLE_INLINE_FUNCTION Real find_min_bounded_val(const Real val1, const Real v
 PORTABLE_INLINE_FUNCTION bool is_near_zero(const Real val, const Real tol) {
   return std::abs(val) < tol;
 }
+} // namespace gruneisen_utils
 /*
 The Gruneisen EOS diverges at a specific compression. Ensure that the maximum density is
 below the smallest singularity in the reference pressure curve
 */
-PORTABLE_FUNCTION Real Gruneisen::ComputeRhoMax(const Real s1, const Real s2,
-                                                const Real s3, const Real rho0) {
+PORTABLE_INLINE_FUNCTION Real Gruneisen::ComputeRhoMax(const Real s1, const Real s2,
+                                                       const Real s3, const Real rho0) {
+  using namespace gruneisen_utils;
+
   // Polynomial from the denominator of the reference pressure curve:
   auto poly = [=](Real eta) { return 1 - s1 * eta - s2 * square(eta) - s3 * cube(eta); };
 
@@ -141,11 +246,13 @@ PORTABLE_FUNCTION Real Gruneisen::ComputeRhoMax(const Real s1, const Real s2,
     return std::numeric_limits<Real>::infinity();
   }
 }
-PORTABLE_FUNCTION Real Gruneisen::Gamma(const Real rho_in) const {
+PORTABLE_INLINE_FUNCTION Real Gruneisen::Gamma(const Real rho_in) const {
   const Real rho = std::min(rho_in, _rho_max);
   return rho < _rho0 ? _G0 : _G0 * _rho0 / rho + _b * (1 - _rho0 / rho);
 }
-PORTABLE_FUNCTION Real Gruneisen::dPres_drho_e(const Real rho_in, const Real sie) const {
+PORTABLE_INLINE_FUNCTION Real Gruneisen::dPres_drho_e(const Real rho_in,
+                                                      const Real sie) const {
+  using namespace gruneisen_utils;
   const Real rho = std::min(rho_in, _rho_max);
   if (rho < _rho0) {
     return square(_C0) + Gamma(rho) * sie;
@@ -164,17 +271,17 @@ PORTABLE_FUNCTION Real Gruneisen::dPres_drho_e(const Real rho_in, const Real sie
            rho * Gamma(rho) * dE_H;
   }
 }
-PORTABLE_FUNCTION Real Gruneisen::InternalEnergyFromDensityTemperature(
+PORTABLE_INLINE_FUNCTION Real Gruneisen::InternalEnergyFromDensityTemperature(
     const Real rho_in, const Real temp, Real *lambda) const {
   return _Cv * (temp - _T0);
 }
-PORTABLE_FUNCTION Real Gruneisen::TemperatureFromDensityInternalEnergy(
+PORTABLE_INLINE_FUNCTION Real Gruneisen::TemperatureFromDensityInternalEnergy(
     const Real rho_in, const Real sie, Real *lambda) const {
   return _T0 + sie / _Cv;
 }
-PORTABLE_FUNCTION Real Gruneisen::PressureFromDensityInternalEnergy(const Real rho_in,
-                                                                    const Real sie,
-                                                                    Real *lambda) const {
+PORTABLE_INLINE_FUNCTION Real Gruneisen::PressureFromDensityInternalEnergy(
+    const Real rho_in, const Real sie, Real *lambda) const {
+  using namespace gruneisen_utils;
   const Real rho = std::min(rho_in, _rho_max);
   Real P_H;
   Real E_H;
@@ -190,12 +297,13 @@ PORTABLE_FUNCTION Real Gruneisen::PressureFromDensityInternalEnergy(const Real r
   }
   return P_H + Gamma(rho) * rho * (sie - E_H);
 }
-PORTABLE_FUNCTION Real Gruneisen::SpecificHeatFromDensityInternalEnergy(
+PORTABLE_INLINE_FUNCTION Real Gruneisen::SpecificHeatFromDensityInternalEnergy(
     const Real rho_in, const Real sie, Real *lambda) const {
   return _Cv;
 }
-PORTABLE_FUNCTION Real Gruneisen::BulkModulusFromDensityInternalEnergy(
+PORTABLE_INLINE_FUNCTION Real Gruneisen::BulkModulusFromDensityInternalEnergy(
     const Real rho_in, const Real sie, Real *lambda) const {
+  using namespace gruneisen_utils;
   const Real rho = std::min(rho_in, _rho_max);
   // The if statement exists here to avoid the divide by zero
   if (rho < _rho0) {
@@ -208,41 +316,38 @@ PORTABLE_FUNCTION Real Gruneisen::BulkModulusFromDensityInternalEnergy(
     return rho * dPdr_e + PressureFromDensityInternalEnergy(rho, sie) / rho * dPde_r;
   }
 }
-PORTABLE_FUNCTION
+PORTABLE_INLINE_FUNCTION
 Real Gruneisen::GruneisenParamFromDensityInternalEnergy(const Real rho_in, const Real sie,
                                                         Real *lambda) const {
   const Real rho = std::min(rho_in, _rho_max);
   return Gamma(rho);
 }
 // Below are "unimplemented" routines
-PORTABLE_FUNCTION Real Gruneisen::PressureFromDensityTemperature(const Real rho_in,
-                                                                 const Real temp,
-                                                                 Real *lambda) const {
+PORTABLE_INLINE_FUNCTION Real Gruneisen::PressureFromDensityTemperature(
+    const Real rho_in, const Real temp, Real *lambda) const {
   const Real rho = std::min(rho_in, _rho_max);
   return PressureFromDensityInternalEnergy(
       rho, InternalEnergyFromDensityTemperature(rho, temp));
 }
-PORTABLE_FUNCTION Real Gruneisen::SpecificHeatFromDensityTemperature(const Real rho_in,
-                                                                     const Real temp,
-                                                                     Real *lambda) const {
+PORTABLE_INLINE_FUNCTION Real Gruneisen::SpecificHeatFromDensityTemperature(
+    const Real rho_in, const Real temp, Real *lambda) const {
   const Real rho = std::min(rho_in, _rho_max);
   return SpecificHeatFromDensityInternalEnergy(
       rho, InternalEnergyFromDensityTemperature(rho, temp));
 }
-PORTABLE_FUNCTION Real Gruneisen::BulkModulusFromDensityTemperature(const Real rho_in,
-                                                                    const Real temp,
-                                                                    Real *lambda) const {
+PORTABLE_INLINE_FUNCTION Real Gruneisen::BulkModulusFromDensityTemperature(
+    const Real rho_in, const Real temp, Real *lambda) const {
   const Real rho = std::min(rho_in, _rho_max);
   return BulkModulusFromDensityInternalEnergy(
       rho, InternalEnergyFromDensityTemperature(rho, temp));
 }
-PORTABLE_FUNCTION
+PORTABLE_INLINE_FUNCTION
 Real Gruneisen::GruneisenParamFromDensityTemperature(const Real rho_in, const Real temp,
                                                      Real *lambda) const {
   const Real rho = std::min(rho_in, _rho_max);
   return Gamma(rho);
 }
-PORTABLE_FUNCTION void Gruneisen::DensityEnergyFromPressureTemperature(
+PORTABLE_INLINE_FUNCTION void Gruneisen::DensityEnergyFromPressureTemperature(
     const Real press, const Real temp, Real *lambda, Real &rho, Real &sie) const {
   sie = _Cv * (temp - _T0);
   // We have a branch at rho0, so we need to decide, based on our pressure, whether we
@@ -266,10 +371,10 @@ PORTABLE_FUNCTION void Gruneisen::DensityEnergyFromPressureTemperature(
     }
   }
 }
-PORTABLE_FUNCTION void Gruneisen::FillEos(Real &rho_in, Real &temp, Real &sie,
-                                          Real &press, Real &cv, Real &bmod,
-                                          const unsigned long output,
-                                          Real *lambda) const {
+PORTABLE_INLINE_FUNCTION void Gruneisen::FillEos(Real &rho_in, Real &temp, Real &sie,
+                                                 Real &press, Real &cv, Real &bmod,
+                                                 const unsigned long output,
+                                                 Real *lambda) const {
   // The following could be sped up with work!
   const unsigned long input = ~output;
   if (thermalqs::temperature & input && thermalqs::pressure & input) {
@@ -291,7 +396,7 @@ PORTABLE_FUNCTION void Gruneisen::FillEos(Real &rho_in, Real &temp, Real &sie,
 }
 
 // TODO(JMM): pre-cache these rather than recomputing them each time
-PORTABLE_FUNCTION
+PORTABLE_INLINE_FUNCTION
 void Gruneisen::ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &press,
                                        Real &cv, Real &bmod, Real &dpde, Real &dvdt,
                                        Real *lambda) const {
@@ -307,3 +412,5 @@ void Gruneisen::ValuesAtReferenceState(Real &rho, Real &temp, Real &sie, Real &p
   dvdt = gm1 * cv / bmod;
 }
 } // namespace singularity
+
+#endif // _SINGULARITY_EOS_EOS_GRUNEISEN_HPP_
